backend/engine/plugins/github_repo_health/cli/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "repo-health-cli"
version = "0.0.1"
dependencies = [
    "jsonschema",
    "octokitpy",
]

[project.scripts]
repo-health = "cli:main"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.black]
line-length = 120
exclude = '/(\.eggs)/'
